+++ src/users/models.py
@@ -1,5 +1,5 @@
 from src.database import Base
-from sqlalchemy import Column, Integer, String, ForeignKey, Date, Computed
+from sqlalchemy import Column, Integer, String, BigInteger
 
 
 class Users(Base):
@@ -9,7 +9,7 @@ class Users(Base):
     login = Column(String, nullable=False)
     fullname = Column(String, nullable=False)
     status = Column(String, nullable=False)
-    aspro_id = Column(Integer, nullable=False)
+    aspro_id = Column(BigInteger, nullable=False)
     hashed_password = Column(String, nullable=False)
 
     class Config:
